Rename writer to formats since the json format can also be read.
Add read_report interface to format objects.
