+++ gcovr/formats/__init__.py
@@ -19,13 +19,13 @@ LOGGER = logging.getLogger("gcovr")
 
 def get_options() -> List[GcovrConfigOption]:
     return [
-        *txt.writer.get_options(),
-        *cobertura.writer.get_options(),
-        *html.writer.get_options(),
-        *json.writer.get_options(),
-        *csv.writer.get_options(),
-        *sonarqube.writer.get_options(),
-        *coveralls.writer.get_options(),
+        *txt.handler.get_options(),
+        *cobertura.handler.get_options(),
+        *html.handler.get_options(),
+        *json.handler.get_options(),
+        *csv.handler.get_options(),
+        *sonarqube.handler.get_options(),
+        *coveralls.handler.get_options(),
     ]
 
 
@@ -43,7 +43,7 @@ def write_reports(covdata: CovData, options):
         generators.append(
             (
                 [options.txt],
-                txt.writer.print_report,
+                txt.handler.write_report,
                 lambda: LOGGER.warning(
                     "Text output skipped - "
                     "consider providing an output file with `--txt=OUTPUT`."
@@ -55,7 +55,7 @@ def write_reports(covdata: CovData, options):
         generators.append(
             (
                 [options.cobertura],
-                cobertura.writer.print_report,
+                cobertura.handler.write_report,
                 lambda: LOGGER.warning(
                     "Cobertura output skipped - "
                     "consider providing an output file with `--cobertura=OUTPUT`."
@@ -67,7 +67,7 @@ def write_reports(covdata: CovData, options):
         generators.append(
             (
                 [options.html, options.html_details, options.html_nested],
-                html.writer.print_report,
+                html.handler.write_report,
                 lambda: LOGGER.warning(
                     "HTML output skipped - "
                     "consider providing an output file with `--html=OUTPUT`."
@@ -79,7 +79,7 @@ def write_reports(covdata: CovData, options):
         generators.append(
             (
                 [options.sonarqube],
-                sonarqube.writer.print_report,
+                sonarqube.handler.write_report,
                 lambda: LOGGER.warning(
                     "Sonarqube output skipped - "
                     "consider providing an output file with `--sonarqube=OUTPUT`."
@@ -91,7 +91,7 @@ def write_reports(covdata: CovData, options):
         generators.append(
             (
                 [options.json],
-                json.writer.print_report,
+                json.handler.write_report,
                 lambda: LOGGER.warning(
                     "JSON output skipped - "
                     "consider providing an output file with `--json=OUTPUT`."
@@ -103,7 +103,7 @@ def write_reports(covdata: CovData, options):
         generators.append(
             (
                 [options.json_summary],
-                json.writer.print_summary_report,
+                json.handler.write_summary_report,
                 lambda: LOGGER.warning(
                     "JSON summary output skipped - "
                     "consider providing an output file with `--json-summary=OUTPUT`."
@@ -115,7 +115,7 @@ def write_reports(covdata: CovData, options):
         generators.append(
             (
                 [options.csv],
-                csv.writer.print_report,
+                csv.handler.write_report,
                 lambda: LOGGER.warning(
                     "CSV output skipped - "
                     "consider providing an output file with `--csv=OUTPUT`."
@@ -127,7 +127,7 @@ def write_reports(covdata: CovData, options):
         generators.append(
             (
                 [options.coveralls],
-                coveralls.writer.print_report,
+                coveralls.handler.write_report,
                 lambda: LOGGER.warning(
                     "Coveralls output skipped - "
                     "consider providing an output file with `--coveralls=OUTPUT`."
@@ -154,7 +154,7 @@ def write_reports(covdata: CovData, options):
             on_no_output()
 
     if not reports_were_written:
-        txt.writer.print_report(
+        txt.handler.write_report(
             covdata, "-" if default_output is None else default_output.abspath, options
         )
         default_output = None
@@ -169,6 +169,6 @@ def write_reports(covdata: CovData, options):
         )
 
     if options.print_summary:
-        txt.writer.print_summary_report(covdata, "-", options)
+        txt.handler.write_summary_report(covdata, "-", options)
 
     return generator_error_occurred
