+++ gcovr/formats/html/templates/directory_page.content.html
