+++ gcovr/formats/html/templates/directory_page.html
