+++ gcovr/formats/html/templates/functions_page.content.html
