+++ gcovr/formats/html/templates/functions_page.html
