+++ gcovr/formats/html/templates/functions_page.summary.html
