+++ gcovr/formats/html/templates/style.details.css
