+++ gcovr/formats/html/templates/style.summary.css
