+++ gcovr/formats/txt/write.py
@@ -43,7 +43,7 @@ MISSING_SEPARATOR = "   "
 LINE_WIDTH = 78
 
 
-def print_report(covdata: CovData, output_file: str, options: Options):
+def write_report(covdata: CovData, output_file: str, options: Options):
     """produce the classic gcovr text report"""
 
     with open_text_for_writing(output_file, "coverage.txt") as fh:
@@ -89,7 +89,7 @@ def print_report(covdata: CovData, output_file: str, options: Options):
         fh.write("-" * LINE_WIDTH + "\n")
 
 
-def print_summary_report(covdata: CovData, output_file: str, options: Options):
+def write_summary_report(covdata: CovData, output_file: str, options: Options):
     """Print a small report to the standard output.
     Output the percentage, covered and total lines and branches.
     """
